+++ src/index.ts
@@ -1,0 +1,9 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+import IReactAISettings from "./IReactAISettings";
+import { reactAI } from "./ReactAI";
+import withAITracking from "./withAITracking";
+
+export { reactAI, IReactAISettings, withAITracking };
+